Various small improvements and compile fixes.

+++ src/ui/src/SetupDialog.cpp
@@ -354,9 +354,9 @@ SetupDialog::SetupDialog (SetupContext* context)
 			feature.second->GetId (), kylaFeatureProperty_Size, &resultSize, &size);
 
 		auto item = new QTreeWidgetItem ();
-		item->setData (0, FeatureSizeRole, QVariant{ size });
+		item->setData (0, FeatureSizeRole, QVariant{ static_cast<qlonglong> (size) });
 		item->setData (0, FeatureDescriptionRole, QVariant{ feature.second->GetDescription ()});
-		item->setData (0, FeatureFeatureIdsIndexRole, QVariant{ featureTreeFeatureIds_.size () });
+		item->setData (0, FeatureFeatureIdsIndexRole, QVariant{ static_cast<qlonglong> (featureTreeFeatureIds_.size ()) });
 		item->setText (0, feature.second->GetTitle ());
 		item->setCheckState (0, Qt::Unchecked);
